pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "domepieces"
version = "0.1.0"
description = ""
authors = ["backwardspy <backwardspy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pycryptodomex = "^3.10.4"
pydantic = "^1.8.2"
SQLAlchemy = "^1.4.25"
plyvel = "^1.3.0"
msgpack = "^1.0.2"

[tool.poetry.group.dev.dependencies]
black = "^21.9-beta.0"
pylint = "^2.11.1"
mypy = "^0.910"
isort = "^5.9.3"
pytest = "^6.2.5"
ptpython = "^3.0.20"
pytest-cov = "^2.12.1"
types-setuptools = "^57.4.0"
types-toml = "^0.10.0"

[tool.isort]
profile = "black"

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = ["sqlalchemy", "plyvel", "msgpack"]
ignore_missing_imports = true

[tool.pylint.message_control]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "too-few-public-methods",
]
good-names = ["db"]
